refactor MediaQuery to be more consistent
