+++ src/@primitives/MediaQuery/enhancer.js
@@ -2,18 +2,7 @@ import { branch, compose, mapProps } from 'recompose';
 import { every } from 'lodash';
 import withWindow from './withWindow';
 import withTheme from '../withTheme';
-
-const queryMatcher = ({ width, height }) => (selector, query) => {
-  switch (query) {
-    case 'maxWidth': return width < selector;
-    case 'minWidth': return width > selector;
-    case 'maxHeight': return height < selector;
-    case 'minHeight': return height > selector;
-    case 'minDeviceAspectRatio': return width / height > selector;
-    case 'maxDeviceAspectRatio': return width / height < selector;
-    default: return true;
-  }
-};
+import queryMatcher from './queryMatcher';
 
 // Make it easy to do breakpoint-based component branching, a-la recompose's branch method.
 // The "test" function should return an object that mirrors what you'd do in a css mediaQuery:
