+++ src/@primitives/MediaQuery/index.js
@@ -1,21 +1,31 @@
 import { Component } from 'react';
 import PropTypes from 'prop-types';
-import { get } from 'lodash';
+import { get, every } from 'lodash';
+import { pick, mapValues, flow, filter } from 'lodash/fp';
 import { compose } from 'recompose';
 
 import withTheme from '../withTheme';
 import withWindow from './withWindow';
+import queryMatcher from './queryMatcher';
 
 export { withWindow };
 export { default as enhancer } from './enhancer';
+
+const supportedMediaQueryTypes = {
+  minWidth: PropTypes.string,
+  maxWidth: PropTypes.string,
+  minHeight: PropTypes.string,
+  maxHeight: PropTypes.string,
+  minDeviceAspectRatio: PropTypes.string,
+  maxDeviceAspectRatio: PropTypes.string,
+};
 
 // <MediaQuery max="md" min="sm">I render on screens sm to md</MediaQuery>
 // <MediaQuery max="xs">I render only on xs screens</MediaQuery>
 // <MeediaQuery min="md">I render on screens md or above</MediaQuery>
 class MediaQuery extends Component {
   static propTypes = {
-    min: PropTypes.string,
-    max: PropTypes.string,
+    ...supportedMediaQueryTypes,
     children: PropTypes.node,
 
     // These props are passed in through HOCs (withTheme and withWindow)
@@ -32,21 +42,17 @@ class MediaQuery extends Component {
   };
 
   static defaultProps = {
-    min: null,
-    max: null,
-    children: null,
+    ...mapValues(supportedMediaQueryTypes, () => null),
   };
 
   get shouldBeVisible() {
-    const {
-      breakpoints, min, max, window: { width },
-    } = this.props;
-    const minSelector = get(breakpoints, min, 0);
-    const maxSelector = get(breakpoints, max, 0);
+    const mediaQuery = flow(
+      pick(Object.keys(supportedMediaQueryTypes)),
+      mapValues(breakpoint => get(this.props.breakpoints, breakpoint)),
+    )(this.props);
 
-    const passesMinWidthRule = minSelector && width > minSelector;
-    const passesMaxWidthRule = maxSelector && width < maxSelector;
-    return passesMaxWidthRule && passesMinWidthRule;
+    const { window: { width, height } } = this.props; // destructuring just to make below line clean
+    return every(mediaQuery, queryMatcher({ width, height }));
   }
 
   render() {
